dir_conf.lua: add a hack for video-aspect
When video-aspect is explicitly set to -1 (default), observing the
property only tells the real aspect ratio for the current video.

This hack attempts to guess, based on video-params/par, if the video
aspect is set to the native value. It will break on videos that
don't have 1:1 pixel to aspect ratio but it's better than the previous
situation where video-aspect would get saved every time the script was
active, wanted or not.